0dte_options_pipeline/requirements.txt
azure-functions==1.21.3
pandas==2.2.1
yfinance==0.2.37
python-dotenv==1.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
pytest==8.0.2
python-dateutil==2.8.2
pytz==2024.1 